requirements.txt
imageio
Pillow
imageio-ffmpeg
omegaconf
einops
wandb[media]
opencv-python
opencv-python-headless
ipdb
tqdm
easydict
scikit-image
numpy
matplotlib
rich
trimesh

torch==2.7.0 --extra-index-url https://download.pytorch.org/whl/cu128
torchvision==0.22.0 --extra-index-url https://download.pytorch.org/whl/cu128

jax[cuda12]
jaxtyping
optax
flax
icecream
lpips_j
clu
scikit-learn
docopt
dm_pix
plyfile
grain
jupyter
natsort